Implement puzzle20
Implemented puzzle20, taking advantage of certain properties of the input data-set. Only including gaps of 3 and 1, and only including at most four gaps of 1 in a row allows handling each possible run of 1gaps size. The 3gaps must be reached, so the number of ways is simply the product of the number of ways between each 3gap, which can be computed from the number of 1gaps between them. Rather than truly checking the input space exhasutively, or even checking all possible sets of gaps between two gaps of 3, I simply checked only those forms that exist in the input, using pre-computed patterns.