Implement puzzle32
Implemented puzzle32 by tracking what each rule can be on the ticket. First all slots that are rendered invalid by values on the tickets are removed, then once a slot and a rule are known paired, that slot is removed from all other rules. This last step is repeated until all pairings are found.