Implement Puzzle35
Implemented Puzzle35 by using an off the shelf parser, by implementing a custom number type that uses the subtract action to instead perform multiplication, meaning addition and multiplication have the same priority. I then replace all '*' with '-' in the input file, and run the parser on each line. To build and execute requires libraries that don't exist in an easy to use way on Windows, so this was built and run in WSL Ubuntu-18.04