Implement puzzle36
Implemented puzzle36 in much the same way as puzzle35, except this time multiplication performs addition and addition performs multiplication, and using three replace statements, all + are replaced with * and vice versa in the input.
This also was built and run under WSL, as the parser used doesn't build nicely under windows.